small changes after linting

+++ include/activation_layer.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "layer.hpp"
+#include <cstddef>
 
 namespace Robbie
 {
@@ -29,7 +30,7 @@ public:
     }
 
     // Return the number of trainable parameters
-    int get_trainable_params() override
+    size_t get_trainable_params() override
     {
         return 0;
     }
