+++ include/dropout_layer.hpp
@@ -71,7 +71,7 @@ public:
     }
 
     // Return the number of trainable parameters
-    int get_trainable_params() override
+    size_t get_trainable_params() override
     {
         return 0;
     }
