+++ include/fc_layer.hpp
@@ -55,7 +55,7 @@ public:
     }
 
     // Return the number of trainable parameters
-    int get_trainable_params() override
+    size_t get_trainable_params() override
     {
         return this->weights.size() + this->bias.size();
     }
